Release v0.7.8
Version normalized as v0.7.8
Parsed from release file e107_v0.7.8_full.tar.gz

+++ e107_admin/admin.php
@@ -11,9 +11,9 @@
 |     GNU General Public License (http://gnu.org).
 |
 |     $Source: /cvsroot/e107/e107_0.7/e107_admin/admin.php,v $
-|     $Revision: 1.36 $
-|     $Date: 2006/11/16 10:41:46 $
-|     $Author: e107coders $
+|     $Revision: 1.39 $
+|     $Date: 2007/01/06 23:09:56 $
+|     $Author: e107steved $
 +----------------------------------------------------------------------------+
 */
 require_once('../class2.php');
@@ -151,7 +151,7 @@ if ('0' == ADMINPERMS) {
 }
 // end auto db update
 
-if (e_QUERY == 'purge') {
+if (e_QUERY == 'purge' && getperms('0')) {
 	$admin_log->purge_log_events(false);
 }
 
@@ -178,15 +178,15 @@ function render_links($link, $title, $description, $perms, $icon = FALSE, $mode 
 			}
 			if ($mode == 'default') {
 				$text .= "<td class='td' style='text-align:left; vertical-align:top; width:20%; white-space:nowrap'
-					onmouseover=\"eover(this, 'forumheader5')\" onmouseout=\"eover(this, 'td')\" onclick=\"document.location.href='".$link."'\">".$icon." ".$tp->toHTML($title,FALSE,"defs emotes_off")."</td>";
+					onmouseover=\"eover(this, 'forumheader5')\" onmouseout=\"eover(this, 'td')\" onclick=\"document.location.href='".$link."'\">".$icon." ".$tp->toHTML($title,FALSE,"defs,emotes_off")."</td>";
 			}
 			else if ($mode == 'classis') {
 				$text .= "<td style='text-align:center; vertical-align:top; width:20%'><a href='".$link."' title='$description'>".$icon."</a><br />
-					<a href='".$link."' title='$description'><b>".$tp->toHTML($title,FALSE,"defs emotes_off")."</b></a><br /><br /></td>";
+					<a href='".$link."' title='$description'><b>".$tp->toHTML($title,FALSE,"defs,emotes_off")."</b></a><br /><br /></td>";
 			}elseif ($mode == 'beginner'){
                 $text .= "<td style='text-align:center; vertical-align:top; width:20%' ><a href='".$link."' >".$icon."</a>
 					<div style='padding:5px'>
-					<a href='".$link."' title='".$description."' style='text-decoration:none'><b>".$tp->toHTML($title,FALSE,"defs emotes_off")."</b></a></div><br /><br /><br /></td>";
+					<a href='".$link."' title='".$description."' style='text-decoration:none'><b>".$tp->toHTML($title,FALSE,"defs,emotes_off")."</b></a></div><br /><br /><br /></td>";
 			}
 			$td++;
 		}
@@ -211,27 +211,30 @@ require_once(e_ADMIN.'includes/'.$pref['adminstyle'].'.php');
 
 function admin_info() {
 	global $tp;
-	$text = "<div style='text-align:center'>
+
+	$width = (getperms('0')) ? "33%" : "50%";
+
+	$ADMIN_INFO_TEMPLATE = "
+	<div style='text-align:center'>
 		<table style='width: 100%; border-collapse:collapse; border-spacing:0px;'>
 		<tr>
-		<td style='width: 33%; vertical-align: top'>";
+			<td style='width: ".$width."; vertical-align: top'>
+			{ADMIN_STATUS}
+			</td>
+			<td style='width:".$width."; vertical-align: top'>
+			{ADMIN_LATEST}
+			</td>";
 
-	$text .= $tp->parseTemplate('{ADMIN_STATUS}');
+    	if(getperms('0'))
+		{
+			$ADMIN_INFO_TEMPLATE .= "
+			<td style='width:".$width."; vertical-align: top'>{ADMIN_LOG}</td>";
+    	}
 
-	$text .= "</td>
-		<td style='width: 33%; vertical-align: top'>";
-
-	$text .= $tp->parseTemplate('{ADMIN_LATEST}');
-
-	$text .= "</td>
-		<td style='width: 33%; vertical-align: top'>";
-
-	$text .= $tp->parseTemplate('{ADMIN_LOG}');
-
-	$text .= "</td>
+   	$ADMIN_INFO_TEMPLATE .= "
 		</tr></table></div>";
 
-	return $text;
+	return $tp->parseTemplate($ADMIN_INFO_TEMPLATE);
 }
 
 function status_request() {
